python/requirements.txt
paho-mqtt==1.5.1
salesforce-api==0.1.40
flask